requirements.txt
# MCP Learning Project Dependencies
# Core MCP and Protocol Support

# MCP Protocol Dependencies
mcp>=1.0.0

# AWS Services
boto3>=1.26.0
botocore>=1.29.0

# Google AI Services
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5

# Database
psycopg2-binary>=2.9.0

# Time and Date
pytz>=2023.3

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: LLM Integration (for advanced features)
langchain>=0.1.0

# Optional: Environment Management
python-dotenv>=1.0.0

# Optional: Logging and Monitoring
structlog>=23.0.0

# Optional: Docker Support
docker>=6.0.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 